pyproject.toml
[tool.poetry]
name = "test-data-star"
version = "0.1.0"
description = ""
authors = ["Michael Berg-Mohnicke <michael.berg@zalf.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
zalfmas-capnp-schemas = "^0.1.27"
hypercorn = "^0.17.3"
pandas = "^2.3.1"
polars = "^1.31.0"
great-tables = "^0.18.0"
python-fasthtml = "^0.12.22"
datastar-py = "^0.6.3"
pycapnp = "^2.0.0"
zalfmas-common = "^0.1.22"
monsterui = "^1.0.21"
aiostream = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"